Rework the handling of outgoing bodies
